правки по структуре и правки в кастормном хуке
