+++ src/components/article-params-form/ArticleParamsForm.module.scss
@@ -5,20 +5,9 @@
 	height: 100%;
 	overflow-y: auto; /* Используем вертикальную прокрутку */
 	background-color: #ffffff; /* Добавляем белый фон для сайдбара */
+	border-right: 1px solid #000000; /* Добавим границу справа от сайдбара */
 	transform: translate(-616px);
 	transition: transform 0.5s ease;
-}
-
-/* Добавим границу справа от сайдбара */
-
-.container::after {
-	content: '';
-	position: absolute;
-	top: 0;
-	right: 0;
-	bottom: 0;
-	width: 1px;
-	background-color: #000000;
 }
 
 .container_open {
@@ -36,7 +25,6 @@
 	min-height: 100%;
 	padding: 50px 31px;
 	background-color: transparent; /* Форма должна быть прозрачной внутри сайдбара */
-	border-right: none; /* Убираем границу у формы */
 }
 
 .bottomContainer {
