+++ src/components/article-params-form/ArticleParamsForm.tsx
@@ -2,7 +2,7 @@ import { ArrowButton } from 'src/ui/arrow-button';
 import { Button } from 'src/ui/button';
 import { RadioGroup } from 'src/ui/radio-group';
 import { Separator } from 'src/ui/separator';
-import { useHandleClickOutside } from './useHandleClickOutside ';
+import { useHandleClickOutside } from 'src/hooks/useHandleClickOutside ';
 
 import styles from './ArticleParamsForm.module.scss';
 import { FormEvent, useRef, useState } from 'react';
@@ -31,45 +31,11 @@ export const ArticleParamsForm = ({ setArticleState }: ArticleParamsProps) => {
 	const [formState, setFormState] = useState(defaultArticleState);
 
 	// Подписываемся на событие клика документа через кастормный хук
-	useHandleClickOutside(isOpen, sidebarRef, () => setIsOpen(false));
-
-	// Подписываемся на событие клика документа
-	// useEffect(() => {
-	// 	if (!isOpen) return; // сразу выходим из useEffect, если сайдбар закрыт
-
-	// 	// Функция для закрытия сайдбара при клике вне его
-	// 	const handleSidebarIfClickedOutside = (e: MouseEvent) => {
-	// 		if (
-	// 			sidebarRef.current &&
-	// 			!sidebarRef.current.contains(e.target as Node)
-	// 		) {
-	// 			setIsOpen(false);
-	// 		}
-	// 	};
-
-	// 	// Добавляем обработчик для закрытия сайдбара по нажатию на клавишу Escape
-	// 	const handleKeyDown = (event: KeyboardEvent) => {
-	// 		if (event.key === 'Escape') {
-	// 			setIsOpen(false);
-	// 		}
-	// 	};
-
-	// 	if (isOpen && sidebarRef.current) {
-	// 		document.addEventListener('mousedown', handleSidebarIfClickedOutside);
-	// 		window.addEventListener('keydown', handleKeyDown); // Добавление обработчика для клавиши Escape
-	// 	}
-
-	// 	// Отписываемся от события при размонтировании компонента
-	// 	return () => {
-	// 		if (sidebarRef.current) {
-	// 			document.removeEventListener(
-	// 				'mousedown',
-	// 				handleSidebarIfClickedOutside
-	// 			);
-	// 			window.removeEventListener('keydown', handleKeyDown); // Удаление обработчика для клавиши
-	// 		}
-	// 	};
-	// }, [isOpen]);
+	useHandleClickOutside({
+		isOpen: isOpen,
+		rootRef: sidebarRef,
+		onClose: () => setIsOpen(false),
+	});
 
 	const handleChangeOption = (
 		field: keyof ArticleStateType,
@@ -86,6 +52,11 @@ export const ArticleParamsForm = ({ setArticleState }: ArticleParamsProps) => {
 		setArticleState(defaultArticleState);
 	};
 
+	const handleSubmit = (e: FormEvent<HTMLFormElement>) => {
+		e.preventDefault();
+		setArticleState(formState);
+	};
+
 	return (
 		<div ref={sidebarRef}>
 			<ArrowButton isOpen={isOpen} onClick={() => setIsOpen(!isOpen)} />
@@ -94,9 +65,8 @@ export const ArticleParamsForm = ({ setArticleState }: ArticleParamsProps) => {
 				<form
 					className={styles.form}
 					onReset={handleReset} // Передача обработчика сброса в onReset
-					onSubmit={(e: FormEvent<HTMLFormElement>) => {
-						e.preventDefault(), setArticleState(formState);
-					}}>
+					onSubmit={handleSubmit} // Передача обработчика применения стилей к странице в onSumbit
+				>
 					<Text as={'h2'} size={31} weight={800} uppercase={true}>
 						Задайте параметры
 					</Text>
